requirements.txt
aiohttp==1.0.5
async-timeout==1.2.1
beautifulsoup4==4.5.3
cffi==1.10.0
chardet==3.0.3
colorama==0.3.9
decorator==4.0.11
discord.py==0.16.7
discord.webhooks==0.0.3
gitdb2==2.0.0
GitPython==2.1.3
lxml==3.7.3
multidict==2.1.5
olefile==0.44
Pillow==4.1.1
prettytable==0.7.2
psutil==5.2.1
pycparser==2.17
PyNaCl==1.0.1
PythonGists==1.1.3
requests==2.12.4
six==1.10.0
smmap2==2.0.1
spice-api==1.0.4
strawpy==1.1
virtualenv==15.1.0
websockets==3.3
win-unicode-console==0.5
